fix: data fetching in static generates sites
