+++ package.json
@@ -64,6 +64,7 @@
     "jiti": "^1.9.1",
     "lodash.groupby": "^4.6.0",
     "marked": "^2.0.3",
+    "mkdirp": "^1.0.4",
     "nuxt-edge": "^2.16.0-26969781.9212d34e",
     "nuxt-extend": "^0.0.3",
     "nuxt-i18n": "^6.24.0",
