+++ src/core/runtime/index.ts
@@ -1,3 +1,4 @@
 export * from './docus'
+export * from './RemoteQueryBuilder'
 export * from './QueryBuilder'
 export * from './utils'
